Fix aa tokenizer.

+++ clasp/utils.py
@@ -28,8 +28,8 @@ def basic_aa_tokenizer(seq, context_length, return_mask=True):
     aa = "ACDEFGHIKLMNOPQRSTUVWY"
     d = {a: i for i, a in enumerate(aa)}
     seq_len = len(seq)
-    seq_empty = torch.zeros(context_length - len(seq))
-    seq_tok   = torch.tensor([d[a] if a in aa else 22 for a in seq])
+    seq_empty = torch.zeros(context_length - len(seq), dtype=torch.long)
+    seq_tok   = torch.tensor([d[a] if a in aa else 22 for a in seq], dtype=torch.long)
     seq = torch.cat([seq_tok, seq_empty], dim=0)#.unsqueeze(0)
     if return_mask:
         mask = torch.zeros_like(seq).bool()
